summarizer_app/requirements.txt
Flask>=2.0
python-dotenv>=1.0
requests>=2.25
beautifulsoup4>=4.9
lxml>=4.6 # Efficient parser for BeautifulSoup
gunicorn>=20.1 # Added for better serving (alternative to flask run)
Werkzeug>=2.0 # For password hashing
Markdown>=3.3 # For rendering summary markdown as HTML
youtube-transcript-api>=0.6 # To fetch YouTube transcripts
openai>=1.0 # Added for OpenAI API support